BUCK files for Android artifacts.
Summary: Ported everything to cxx_library and added target for djinni bindings, will add targets for Java part in a separate diff. Will commit it once I get all independent diffs working together.

Test Plan: `buck build native/third-party/rocketspeed/github:rocketbench#android-arm`

Reviewers: pja, ranji42, dhruba

Reviewed By: dhruba

Subscribers: rocketspeed

Differential Revision: https://reviews.facebook.net/D33135

+++ src/messages/descriptor_event.cc
@@ -5,6 +5,10 @@
 //
 #include "descriptor_event.h"
 
+#include <cstddef>
+#include <cerrno>
+#include <cstring>
+
 namespace rocketspeed {
 
 DescriptorEvent::DescriptorEvent(std::shared_ptr<Logger> info_log, int fd)
@@ -32,10 +36,10 @@ Status DescriptorEvent::Write(std::string&& data) {
     return Status::IOError(strerror(errno));
   } else if (static_cast<size_t>(count) != data.size()) {
     LOG_WARN(info_log_,
-             "Wanted to write %zu bytes to fd(%d) but written (%zd).",
+             "Wanted to write %zu bytes to fd(%d) but written (%zu).",
              data.size(),
              fd_,
-             count);
+             static_cast<size_t>(count));
     return Status::IOError("Partial write.");
   }
   return Status::OK();
